:sparkles: feat: requisições funcionando como o esperado
