+++ backend/src/database/controllers/LoginController.js
@@ -1,8 +1,8 @@
-const UserService = require('../services/UserService');
+const LoginService = require('../services/LoginService');
 
 const login = async (req, res) => {
   const { email, password } = req.body;
-  const user = await UserService.login(email, password)
+  const user = await LoginService.login(email, password)
   return res.status(user.status).json(user.message);
 }
